Weighted sampling and bulk-add examples
Weighted sampling has been added and is also the new default. With weighted sampling enabled, shorter sentences are preferred over longer ones.

Bulk-add option has been added to permanently write example sentences to the destination field. This is especially useful for mobile users of e.g. AnkiDroid. The size increase is limited; in my experience about 1kB per note. The number of sentences to be permanently written to the destination field ("MAX_PERMANENT") is set to 5 by default.

When rehearsing when this add-on is loaded (i.e. desktop or laptop), the add-on will temporarily overwrite to the destination field. The idea behind this is that the user will get exposed to different sentences when the computational cost is low.