Do not start the list of examples with a newline
The list of examples was started, and ended, with a newline. Although it may affect formatting of existing users, this is a undesired behavior (and a remainer of the past). Should users want the extra newline separating examples from the rest of their fields, they can edit their templates accordingly.

Closes #2